Address feedback, fix some issues, update test

+++ src/coreclr/nativeaot/System.Private.CoreLib/src/System/Threading/Condition.cs
@@ -107,7 +107,7 @@ namespace System.Threading
             Waiter waiter = GetWaiterForCurrentThread();
             AddWaiter(waiter);
 
-            uint recursionCount = _lock.ExitAll();
+            ushort recursionCount = _lock.ExitAll();
             bool success = false;
             try
             {
